requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-qdrant
langchain_openai
pandas
python-dotenv
qdrant-client
azure-storage-blob
azure-ai-textanalytics
